back-end/requirements.txt
alembic==1.7.5
click==8.0.3
coverage==6.2
Flask==2.0.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
PyJWT==2.3.0
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.29
Werkzeug==2.0.2
